Add files via upload
WHILE LOOP:
A while loop allows your code to repeat itself based on a condition you set.
It is similar to an if statement in that you ask a question, and as long as the answer is true, the computer will repeatedly run the code.

REMEMBER TWO THINGS:
a). Change the variable that control the condition within the loop itself.
b). You have to be really careful that you don't accidentally invoke an infinite loop! This is where the computer will loop code until the end of time. Without a break. Forever. 😭