Gen-Identifier
Based on concepts of Casting ---
CASTING
If statements support more than ==. They support inequality symbols as well. Remember those <, >, and = signs you used in math way back when? Well they are back again, but this time they look a bit different.

# equal
5 == 5

# not equal
3 != 5

# greater than
5 > 3

# greater than or equal to
5 >= 3

# less than
3 < 5

# less than or equal to
3 <= 5

Casting is where we explicitly tell the computer that what we are typing is a number and not a letter.

CHALLENGE FOR DAY 9:
Generation Generator
What generation are you a part of?

Use this list to guide you.

Generation Name
Traditionalists	1925-1946
Baby Boomers	1947-1964
Generation X	1965-1981
Millennials	1982-1995
Generation Z	1996-2015