Use of 'continue' command
So far we've used break in the while True loop. break leaves the loop completely and runs the next unindented line of code. However, you may want to stop the code and start the loop from the top again. (This is ideal for building games!)

The continue command stops executing code in the loop and starts at the top of the loop again. Essentially, we want to kick the user back to the original question.
Adding continue will start the code from the start and ask the first question again.

exit():
The exit() function completely stops the program and it will not run any more lines of code.