Subroutine: The Recipe for Coding
𝐒𝐮𝐛𝐫𝐨𝐮𝐭𝐢𝐧𝐞𝐬
So far, when we wanted to repeat code, we have had to use loops or copy and paste code.

What if I told you there was a way to use code or call it and use it anytime anywhere??

That is a subroutine!!

A subroutine tells the computer that a piece of code exists and to go run that code again and again...

Just like a recipe
Subroutines work like a recipe in a cookbook. If you want to know how to make a cake, you don't have to start from scratch every time. You can use a recipe to get the same quality each time.

How do we define a subroutine?
A subroutine is defined by:
𝒅𝒆𝒇 (𝒘𝒉𝒊𝒄𝒉 𝒔𝒕𝒂𝒏𝒅𝒔 𝒇𝒐𝒓 𝒅𝒆𝒇𝒊𝒏𝒊𝒕𝒊𝒐𝒏)

You need to add () even if there are no arguments followed by a colon :. The code needs to be indented.

𝗟𝗲𝘁'𝘀 𝘁𝗿𝘆 𝗶𝘁
👉 Let's roll a random number on a six-sided dice:
def rollDice():
  import random
  dice = random.randint(1, 6)
  print("You rolled", dice)

Why is nothing happening??

In this code, I have defined how to roll a dice (this is my recipe for rolling a dice), but I have not actually 'called' the program to run.

𝐂𝐚𝐥𝐥 𝐭𝐡𝐞 𝐒𝐮𝐛𝐫𝐨𝐮𝐭𝐢𝐧𝐞
Calling the subroutine means telling it to run.

👉 We need to 'call' the code by adding one more line to our code with the name of the subroutine and the empty ():
def rollDice():
  import random
  dice = random.randint(1, 6)
  print("You rolled", dice)

rollDice()

Note that when you call the subroutine, you need to ensure you do NOT indent.

I can even add a range and roll the dice 100 times by adding this code:
for i in range(100):
  rollDice()