Parameters in Subroutines
𝗣𝗮𝗿𝗮𝗺𝗲𝘁𝗲𝗿𝘀

I get it. So far, subroutines have been a bit underwhelming...

Let's put those subroutines to better use by sending them information using parameters and making them do different things based on different inputs.

If you change the ingredients in a recipe, you get a different kind of cake. Let's do that with subroutines.

In a subroutine, the () are for the argument (FYI argument is another word for parameter). These are the pieces of information we pass to the code. These can be variable names that are made up for the first time within the argument ().

Here is a simple subroutine that uses the argument to take in the name of an ingredient and expresses its opinion (quite strongly) about the ingredient that the user typed. For example, 'chocolate' is amazing, but 'broccoli'...not so much.

👉 Try it out.
def whichCake(ingredient):
  if ingredient == "chocolate":
    print("Mmm, chocolate cake is amazing")
  elif ingredient == "broccoli":
    print("You what mate?!")
  else: 
    print("Yeah, that's great I suppose...")


𝑩𝒖𝒕 𝒉𝒐𝒘 𝒅𝒐 𝒘𝒆 𝒄𝒂𝒍𝒍 𝒊𝒕?

𝗛𝗼𝘄 𝗱𝗼 𝘄𝗲 𝗰𝗮𝗹𝗹 𝘁𝗵𝗲 𝘀𝘂𝗯𝗿𝗼𝘂𝘁𝗶𝗻𝗲?

We call it in the same way as before. However, instead of leaving the () blank, we send the code a message.

👉 What happens when you add this to the end of your code above?
whichCake("chocolate")

𝗔𝗱𝗱𝗶𝗻𝗴 𝗠𝗼𝗿𝗲 𝗔𝗿𝗴𝘂𝗺𝗲𝗻𝘁𝘀
We can have as many arguments as we want, separated by commas.

Now, this subroutine is expecting three arguments: ingredient, base, and coating:
def whichCake(ingredient, base, coating)

👉 Let's update our code to now show all three arguments:
def whichCake(ingredient, base, coating):
  if ingredient == "chocolate":
    print("Mmm, chocolate cake is amazing")
  elif ingredient == "broccoli":
    print("You what mate?!")
  else: 
    print("Yeah, that's great I suppose...")
  print("So you want a", ingredient, "cake on a", base, "base with", coating, "on top?")

whichCake("carrot", "biscuit", "icing")

👉 I could even ask the user to name an ingredient, base, and coating by adding:
def whichCake(ingredient, base, coating):
  if ingredient == "chocolate":
    print("Mmm, chocolate cake is amazing")
  elif ingredient == "broccoli":
    print("You what mate?!")
  else: 
    print("Yeah, that's great I suppose...")
  print("So you want a", ingredient, "cake on a", base, "base with", coating, "on top?")

userIngredient = input("Name an ingredient: ")
userBase = input("Name a type of base: ")
userCoating = input("Fave cake topping: ")
whichCake(userIngredient, userBase, userCoating)