👉 Day 24 Challenge: Infinity Dice 🎲
👉 Day 24 Challenge
Let's build an infinity dice!

Create subroutines that will roll a dice with any number of sides (essentially as big of a number as you like). Write one subroutine with one parameter that allows us to call a function (such as rollDice).

Example:

Infinity Dice 🎲

How many sides?: 600
You rolled 532

Roll again? yes
You rolled 102

Roll again? no

𝗛𝗶𝗻𝘁𝘀 💡:
a. Import your library first.
b. Create a variable to ask the user how many sides the dice should be.
c. Create a variable that allows the user to leave the game.
d. Create a subroutine with rollDice() that sets the sides of the dice as the argument.
e. Create a while loop that allows the user to roll again or leave the game.