👉 Day 25 Challenge: ⚔️Character Stats Generator⚔️
👉 Day 25 Challenge
Let's extend Day 24's project and create a health stats generator for a character in a video game.

Create a subroutine that rolls a dice of any size and returns that number.
Then, create a second subroutine that rolls one six-sided dice and one eight-sided dice.
Multiply the number from the six-sided dice and eight-sided dice together and return that subroutine as a character's health stats for a video game.
Print out the character's health stats.
Add a loop to give the user the option to generate health stats for another character.
(We genuinely see this in video games!)

🥳 Extra points if you ask for the character's name and double extra points if you use different colors!

⚔️ Character Stats Generator ⚔️

Name your warrior: Agnes
Their health is: 20hp

Name your warrior: Ian
Their health is: 6hp

Name your warrior: Penelope
Their health is: 48hp


𝗛𝗜𝗡𝗧𝗦 💡:
a. Import your library first.
b. Create one subroutine for a dice of any number and return it.
c. Create a subroutine that rolls a dice with numbers 1-6 and a dice with numbers 1-8 and multiplies the two numbers together. Return that subroutine.
d. Ask the user to name their character and print that character's health stats.
e. Create a while loop to allow the user to generate a new character's health stats.