👉 Day 29 Challenge: Super Subroutine for Colors
1. Write a subroutine that writes text in color. All it will do is print out the text in that color and turn the color back to normal when it's finished.
2. Control end and sep so there are not random symbols or spaces.

𝗛𝗜𝗡𝗧𝗦 💡 ::
a. Create a subroutine with a variable for 'word' and 'color' through the use of if statements. Here is an example of how to include the color red:
[ ]:  if color=="red":
[ ]:      print("\033[31m", word, sep="", end="")
b. print your title 'Super Subroutine' and 'with my' as regular print statements.
c. print the rest of the text in different colors by calling the subroutine.