Day 32: Lists: List Modification and Loops
𝗖𝗵𝗮𝗻𝗴𝗶𝗻𝗴 𝗟𝗶𝘀𝘁𝘀
We can also change lists and the index.

We built our list with timetable =, but we want to change index 4, "sport". We can do this by calling it with [ ].

[ ]:  timetable = ["Computer Science", "Math", "English", "Art", "Sport"]
[ ]:  print(timetable[0])
[ ]:  print(timetable[1])
[ ]:  print(timetable[2])
[ ]:  print(timetable[3])
[ ]:  print(timetable[4])

👉 Add this to to the code above:
[ ]:  timetable[4]= "Watch TV"

Why is it not printing correctly? I have created the timetable, printed it out, and changed index 4 of the timetable.

👉 However, I need to print the changed version. Let's print our new index:
[ ]:  timetable[4]= "Watch TV"
[ ]:  print(timetable[0])
[ ]:  print(timetable[1])
[ ]:  print(timetable[2])
[ ]:  print(timetable[3])
[ ]:  print(timetable[4])


𝗟𝗶𝘀𝘁𝘀 𝗮𝗻𝗱 𝗟𝗼𝗼𝗽𝘀

Why would I want to write all of those lines of code?

Introducing lists' best friend...loops
👉 We can replace a lot of those lines of code we just wrote with just two lines of code. Change your code to look like this:
[ ]:  timetable = ["Computer Science", "Math", "English", "Art", "Watch TV"]
[ ]:  for lesson in timetable:
[ ]:    print(lesson)

Now we have given the computer a list. We said "make lesson each value within this list and then do something with it."

𝙁𝙡𝙖𝙨𝙝𝙗𝙖𝙘𝙠 ▼
Remember, for loops work by creating the variable right after the word for and setting it equal to each value in a list (so far we have only used numbers with the range function).