👉 Day 33 Challenge: To-Do List
Create your own to do list manager. (This can be super useful!)

1. Ask the user whether they want to view, add, or edit their to do list.
2. If they want to view it, print it out in a nice way (Hint: subroutine).
3. If they choose to add an item to the to do list, allow them to type in the item and then add it to the bottom of the list.
4. If they want to edit the to do list, ask them which item they completed, and remove it from the list.
5. Don't worry about duplicates!
6. The first item you put in the list should be the first item you remove.
7. Add a title, some color, alignment to the text, or emojis. Show off your skills!

EXAMPLE:
To Do List Manager:
Do you want to view, add, or edit your to do list?
view

Record video for day 34

𝗛𝗜𝗡𝗧𝗦 💡::
a. Import libraries first. What libraries will allow you to clear your code and pause it?
b. Remember, your list name needs to come before any subroutines or lists.
c. Create a subroutine very similar to the one we looked at in the common errors section.
d. Add a while True loop with if statements that allow the user to add, edit, or view their to do list.
e. Also, remember what we did to fix the second common error related to removing an item that is not in the to do list. (a nested if statement)
f. End with pausing and clearing your code.