Day 34: 𝗣𝗿𝗲𝘁𝘁𝘆 𝗣𝗿𝗶𝗻𝘁: Using the Index
𝐔𝐬𝐢𝐧𝐠 𝐭𝐡𝐞 𝐈𝐧𝐝𝐞𝐱

We can also use loops in a slightly different way to access elements in a list (or array). With the way we have our code written right now, our list creates a variable called email, sets it equal to the first value (with our counter we set 1 as the first value), and then counts on as we go. (2, 3, 4, etc.)

However, we can also set the for loop to use the index. What should happen is index will start at 0, go through the entire code in the loop at 0, go back to the start of the loop, add 1, and then loop again and again until the list ends.....

👉 Let's tweak the for loop in our subroutine just a bit to accomplish this:
[ ]:  import os, time
[ ]:  listOfEmail = []

[ ]:  def prettyPrint():
[ ]:    os.system("clear") 
[ ]:    print("listofEmail") 
[ ]:    print()
[ ]:    for index in range(len(listOfEmail)): # len counts how many items in a list
[ ]:      print(f"{index}: {listOfEmail[index]}") 
[ ]:    time.sleep(1)


[ ]:  while True:
[ ]:    print("SPAMMER Inc.")
[ ]:    menu = input("1. Add email\n2: Remove email\n3. Show emails\n4. Get SPAMMING\n> ")
[ ]:    if menu == "1":
[ ]:      email = input("Email > ")
[ ]:      listOfEmail.append(email)
[ ]:    elif menu =="2":
[ ]:      email = input ("Email > ")
[ ]:      if email in listOfEmail:
[ ]:        listOfEmail.remove(email)
[ ]:    elif menu == "3":
[ ]:      prettyPrint() 
[ ]:    time.sleep(1)
[ ]:    os.system("clear")


𝒍𝒆𝒏 counts how many items are in a list. In this case, it is starting at 0 and then keeps going until it reaches the end of our data inside our list. We can now eliminate the counter variable because we are counting with index.

We do need to access email in a different way now. We are going to use the actual list name itself and then [𝒊𝒏𝒅𝒆𝒙] to access the 𝒊𝒏𝒅𝒆𝒙.