Day 36: 𝗦𝘁𝗿𝗶𝗻𝗴 𝗠𝗮𝗻𝗶𝗽𝘂𝗹𝗮𝘁𝗶𝗼𝗻
𝐒𝐭𝐫𝐢𝐧𝐠 𝐌𝐚𝐧𝐢𝐩𝐮𝐥𝐚𝐭𝐢𝐨𝐧

Let's do some string manipulation to make if statements even easier.

👉 Does this code look familiar from the insult generator project?
[ ]:  name = input("What's your name? ")
[ ]:  if name == "David" or name == "david":
[ ]:    print("Hello Baldy!")
[ ]:  else: 
[ ]:    print("What a beautiful head of hair!")

Right now, if the user writes "DAVID" or "david", the 🅸🅵 statement works correctly. However, "DaVID" does not give the correct output.

To the computer, " david", "dAviD", and "david" are completely different.

To simplify what the user typed in, we can add these functions to the end of the name of the variable:
▸ .lower() = all letters are lower case
▸ .upper() = all letters are upper case
▸ .title() = capital letter for the first letter of every word
▸ .capitalize() = capital letter for the first letter of only the first word


𝗹𝗼𝘄𝗲𝗿
👉 The computer is converting everything to lowercase before it compares my if statements.

You need to type your if statement in lower case when you use .lower. The if statement needs to be written in upper case when you use .upper, etc.
[ ]:  name = input("What's your name? ")
[ ]:  if name.lower() == "david": 
[ ]:    print("Hello Baldy!")
[ ]:  else: 
[ ]:    print("What a beautiful head of hair!")

𝗪𝗵𝗮𝘁 𝗶𝗳 𝘄𝗲 𝗽𝘂𝘁 𝗮 𝘀𝗽𝗮𝗰𝗲 𝗳𝗶𝗿𝘀𝘁?

Adding .strip() removes any spaces on either side of the word.

👉 We can chain these functions together.
[ ]:  name = input("What's your name? ")
[ ]:  if name.lower().strip() == "david": 
[ ]:    print("Hello Baldy!")
[ ]:  else: 
[ ]:    print("What a beautiful head of hair!")

​🇹​​🇷​​🇾​ ​🇮​​🇹​ ​🇴​​🇺​​🇹​