Day 39: 𝗛𝗮𝗻𝗴𝗺𝗮𝗻
𝐇𝐚𝐧𝐠𝐦𝐚𝐧

This simple game is actually pretty complex. This hangman project combines lists, string manipulation, and slicing, along with other concepts.

It's a bit of a step up, so don't be afraid to check hints and solutions if you're getting stuck. Also, check out our 100 Days of Code forum (https://replit.com/community-hub) for help.

𝗣𝗶𝗰𝗸𝗶𝗻𝗴 𝗳𝗿𝗼𝗺 𝗮 𝗹𝗶𝘀𝘁 𝗿𝗮𝗻𝗱𝗼𝗺𝗹𝘆
👉 random.choice() picks a random item from a list.

Here's a list of words:

listOfWords = ["british", "suave", "integrity", "accent", "evil", "genius", "Downton"]

To select randomly from this list, we assign random.choice() to a variable. We give random.choice() the name of the list in the brackets.

[ ]:  listOfWords = ["british", "suave", "integrity", "accent", "evil", "genius", "Downton"]
[ ]:  wordChosen = random.choice(listOfWords)

There's lots of other stuff needed to make this program work, but, we'll break that down in the challenge section of today's tutorial.

Try it out and get randomizing!