Day 40: 𝗗𝗶𝗰𝘁𝗶𝗼𝗻𝗮𝗿𝗶𝗲𝘀: Common Errors
𝗖𝗼𝗺𝗺𝗼𝗻 𝗘𝗿𝗿𝗼𝗿𝘀

1). 𝗣𝗿𝗶𝗻𝘁𝗶𝗻𝗴 𝘄𝗶𝘁𝗵 𝗸𝗲𝘆𝘀 𝗮𝗻𝗱 𝗳𝗦𝘁𝗿𝗶𝗻𝗴𝘀
👉 Let's run this code and see what happens:
[ ]:  myUser = {"name": "David", "age": 128}

[ ]:  print(f"Your name is {myUser["name"]} and your age is {myUser["age"]}")

ANSWER ---
Note that we have to put the keys in single quotation marks '' inside the fString when using this technique.

This is because we've already used double quotes to start and end the fString. So, using "" for the dictionary value would get Python all confuzzled.
[ ]:  myUser = {"name": "David", "age": 128}

[ ]:  print(f"Your name is {myUser['name']} and your age is {myUser['age']}")

[ ]:  # This code outputs 'Your name is David and your age is 128'.


2). 𝗦𝘆𝗻𝘁𝗮𝘅 𝗲𝗿𝗿𝗼𝗿?
👉 Why are you getting a syntax error on the print statement line?
[ ]:  myUser = {"name": "David", "age": 128}
[ ]:  print(myUser{"name"})

ANSWER ---
The print statement uses square brackets. Curly braces {} are only used to call the value.
[ ]:  myUser = {"name": "David", "age": 128}
[ ]:  print(myUser["name"])


3). 𝗨𝗻𝗱𝗲𝗳𝗶𝗻𝗲𝗱?
👉 What is the problem here?
[ ]:  myUser = {name: "David", "age": 128}
[ ]:  print(myUser["name"])

ANSWER ---
The key, name, in the dictionary should be inside quotes.
[ ]:  myUser = {"name": "David", "age": 128}
[ ]:  print(myUser["name"])


4). 𝗦𝗽𝗮𝗿𝗲 𝗞𝗲𝘆?
👉 What is the problem here?
[ ]:  myUser = {name:"David", "age": 128, "age" = 129}
[ ]:  print(myUser)

ANSWER ---
A dictionary can't have two keys with the same name. It always overrides the previous one. Therefore, the 129 overrides the age, 128.
[ ]:  myUser = {name:"David", "age": 128}

[ ]:  myUser["age"] = 129

[ ]:  print(myUser)