Day 43: Take Your List To A New Dimension
Pay close attention, folks, because 2D lists are basically tables.

Tables are two-dimensional data structures where we can store data both vertically and horizontally.

Usually this means that vertical data is used for fields (one category - name, ID, favorite biscuit, etc.) and horizontal data is used for records (all the data for each category).

Behind the scenes, we see a list inside a list. Forget what you know about reading a table with math or geography: 'across the corridor (x-axis) first and then down the stairs (y-axis)'.

Here, we will do row index first and then the column index.

𝗥𝗲𝗺𝗲𝗺𝗯𝗲𝗿...
Here's a 1D list. We have the list name as a variable, single equals to set the value, and square brackets to show this is a list. 2D lists are very similar.
[ ]:  my1DList = ["Johnny", 21, "Mac"]

𝗔𝗱𝗱𝗶𝗻𝗴 𝗧𝗵𝗲 𝗦𝗲𝗰𝗼𝗻𝗱 𝗗𝗶𝗺𝗲𝗻𝘀𝗶𝗼𝗻
To add the second dimension, we put lists inside the first list.

👉 Each new list has its own set of square brackets and is separated by a comma. This layout of code is nice to help us visualize the 2D list a a table, but...
[ ]:  my2DList = [ ["Johnny", 21, "Mac"],
[ ]:               ["Sian", 19, "PC"],
[ ]:               ["Gethin", 17, "PC"] ]

...you can also lay it out like this below:

[ ]:  my2DList = [ ["Johnny", 21, "Mac"], ["Sian", 19, "PC"], ["Gethin", 17, "PC"] ]

Now, let's print this list.

𝗣𝗿𝗶𝗻𝘁𝗶𝗻𝗴 𝗙𝗿𝗼𝗺 𝗮 𝟮𝗗 𝗟𝗶𝘀𝘁
𝙩𝙝𝙚 𝙚𝙣𝙩𝙞𝙧𝙚 𝙡𝙞𝙨𝙩
👉 We can print an entire 2D list just like we do with a 1D list. However, this will output (print) all of the square brackets, commas, etc.
[ ]:  print(my2DList)

[ ]:  # This will output [['Johnny', 21, 'Mac'], ['Sian', 19, 'PC'], ['Gethin', 17, 'PC']]

𝙖 𝙨𝙞𝙣𝙜𝙡𝙚 𝙧𝙤𝙬
To print a single row, use a single square bracket [] in the print command. However, you will still get all of those square brackets and commas.

👉 In this example, I'm outputting the first row (index 0) - all of the data about Johnny.
[ ]:  print(my2DList[0])

[ ]:  # This code outputs ['Johnny', 21, 'Mac'].

𝙖 𝙨𝙞𝙣𝙜𝙡𝙚 𝙥𝙞𝙚𝙘𝙚 𝙤𝙛 𝙙𝙖𝙩𝙖
The first square bracket references the list, while the second references the item inside that list.

👉 Here are a couple of examples:
[ ]:  my2DList = [ ["Johnny", 21, "Mac"],
[ ]:               ["Sian", 19, "PC"],
[ ]:               ["Gethin", 17, "PC"] ]

[ ]:  print(my2DList[0][0])
[ ]:  # This code outputs 'Johnny'. It's Johnny's name from list 0 (first square bracket), item 0 (second square bracket)

[ ]:  print(my2DList[1][2])
[ ]:  # This code outputs 'PC'. It's Sian's computing preference from list 1 (first square bracket), item 2 (second square bracket)