Day 43: 𝟮𝗗 𝗟𝗶𝘀𝘁: Mecca Your Nan Very Happy
👉 Mecca Your Nan Very Happy
Today's challenge is to create a bingo card. Oh yes, because programming's not just for you hip, young cats. 😆

Anyway, your challenge is to enable "gambling for the elderly" (aka Bingo), and you'll achieve it like this:

1. Randomly generate a series of number between 0 and 90.
2. Allocate each number to a place in a 2D list.
3. The numbers should be in numerical order, left to right.
4. Numbers should not be repeated.
5. The center square should not contain a number. It should contain the word 'BINGO!'.
6. When the program is run, the bingo card should be displayed on screen.


𝗛𝗜𝗡𝗧𝗦 💡::
a. Make sure you include 'prettyprinting'.
b. Try using a 2D list with each sub list as a row.
c. Randomly generate the numbers and append each to a list as you do.
d. Use .sort() to put the list of numbers in order before adding to the card.