👉 Day 44 Challenge: Bingo and Number Guess
👉 𝗗𝗮𝘆 𝟰𝟰 𝗖𝗵𝗮𝗹𝗹𝗲𝗻𝗴𝗲

It's time for more bingo! You can reuse your code from day 43, but this time add the following features:

1. Repeatedly ask the user what number comes up next.
2. Check the bingo card to see if the number picked matches one on the card.
3. If the bingo card is all 'X's, then the user has won.


𝗛𝗜𝗡𝗧𝗦 💡::
a. Create a subroutine called createCard to clean up some of the code from Day 43.
b. Use a variable and loop to store how many x's there are on the card. Add one every time a number is replaced.
c. Check the variable every time to see if it has reached the magic winning number.